requirements.txt
pycryptodome==3.23.0
PyQt6==6.9.1
pyqt6_sip==13.10.2
Requests==2.33.1
